server/requirements.txt
numpy==1.22.3
pymongo==3.11.3
sentence_transformers==2.2.0
fastapi==0.75.2
python-dotenv==0.20.0
scikit_learn==1.0.2
